+++ internal/tfutils/certificate_generator.go
@@ -1,4 +1,4 @@
-package provider
+package tfutils
 
 import (
 	"crypto/rand"
